# static/Api.yaml
openapi: 3.0.0
info:
  title: Onderzoek API
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /api/register_organisatie:
    post:
      summary: "Registreer een organisatie"
      operationId: "registerOrganisatie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                naam:
                  type: string
                email:
                  type: string
                telefoonnummer:
                  type: string
                contactpersoon:
                  type: string
                beschrijving:
                  type: string
                website:
                  type: string
                adres:
                  type: string
              required:
                - naam
                - email
                - telefoonnummer
                - contactpersoon
                - beschrijving
                - website
                - adres
            examples:
              example1:
                summary: "Voorbeeld registratie"
                value:
                  naam: "Api test A"
                  email: "contact@orgorg.com"
                  telefoonnummer: "316463488"
                  contactpersoon: "Rob Jansen"
                  beschrijving: "Dit is een voorbeeldorganisatie"
                  website: "www.organisatie3.com"
                  adres: "Straatnaam 3, 1354 AB, Rotterdam"
      responses:
        '200':
          description: "Organisatie succesvol geregistreerd"
          content:
            application/json:
              examples:
                example1:
                  summary: "Succesvolle registratie response"
                  value:
                    api_key: "GVTsTGyvgtyozZfTWs-FG3d5DOm--xbsS4tGCE6hhss"
                    message: "Organisatie succesvol geregistreerd!"
                    note: "Bewaar deze API key goed, je hebt deze nodig voor toekomstige API-aanvragen."
      security:
        - ApiKeyAuth: []

  /update_onderzoeksvraag/{id}:
    patch:
      summary: "Onderzoeksvraag aanpassen"
      operationId: "updateOnderzoek"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                titel:
                  type: string
                beschrijving:
                  type: string
                plaats:
                  type: string
                max_deelnemers:
                  type: integer
                min_leeftijd:
                  type: integer
                max_leeftijd:
                  type: integer
                beloning:
                  type: string
                datum:
                  type: string
                  format: date
                datum_tot:
                  type: string
                  format: date
                begeleiders:
                  type: string
              required:
                - titel
                - beschrijving
                - plaats
                - max_deelnemers
                - min_leeftijd
                - max_leeftijd
                - beloning
                - datum
                - datum_tot
                - begeleiders
            examples:
              example1:
                summary: "Voorbeeld update onderzoek"
                value:
                  titel: "Test Onderzoek"
                  beschrijving: "Dit is een testonderzoek"
                  plaats: "Assen"
                  max_deelnemers: 15
                  min_leeftijd: 50
                  max_leeftijd: 65
                  beloning: "Ja"
                  datum: "2025-09-01"
                  datum_tot: "2025-10-01"
                  begeleiders: "Nee"
      responses:
        '200':
          description: "Onderzoeksvraag succesvol bijgewerkt"
          content:
            application/json:
              examples:
                example1:
                  summary: "Succesvolle update response"
                  value:
                    message: "Onderzoeksvraag succesvol bijgewerkt"
      security:
        - ApiKeyAuth: []

  /add_onderzoek:
    post:
      summary: "Nieuw onderzoek toevoegen"
      operationId: "addOnderzoek"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type_onderzoek:
                  type: integer
                onderzoekstitel:
                  type: string
                omschrijving:
                  type: string
                plaats:
                  type: string
                aantal_deelnemers:
                  type: integer
                min_leeftijd:
                  type: integer
                max_leeftijd:
                  type: integer
                beperking:
                  type: integer
                begeleiders:
                  type: string
                startdatum:
                  type: string
                  format: date
                einddatum:
                  type: string
                  format: date
                beloning:
                  type: string
              required:
                - type_onderzoek
                - onderzoekstitel
                - omschrijving
                - plaats
                - aantal_deelnemers
                - min_leeftijd
                - max_leeftijd
                - beperking
                - begeleiders
                - startdatum
                - einddatum
                - beloning
            examples:
              example1:
                summary: "Voorbeeld onderzoek toevoegen"
                value:
                  type_onderzoek: 1
                  onderzoekstitel: "Onderzoek via API"
                  omschrijving: "Test via Bruno"
                  plaats: "Utrecht"
                  aantal_deelnemers: 10
                  min_leeftijd: 18
                  max_leeftijd: 65
                  beperking: 2
                  begeleiders: "Nee"
                  startdatum: "2025-04-01"
                  einddatum: "2025-06-01"
                  beloning: "Ja"
      responses:
        '200':
          description: "Onderzoek succesvol toegevoegd"
          content:
            application/json:
              examples:
                example1:
                  summary: "Succesvolle toevoeging response"
                  value:
                    message: "Onderzoek succesvol toegevoegd"
      security:
        - ApiKeyAuth: []

  /api/get-onderzoeken:
    get:
      summary: "Alle onderzoeken ophalen"
      description: "Retourneert alle onderzoeken die behoren tot de organisatie, geïdentificeerd via de API-sleutel in de request headers."
      parameters:
        - name: API-Key
          in: header
          description: "De API-sleutel die aan de organisatie is gekoppeld."
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Lijst met onderzoeken succesvol opgehaald"
          content:
            application/json:
              schema:
                type: object
                properties:
                  onderzoeken:
                    type: array
                    items:
                      $ref: '#/components/schemas/Onderzoeksvraag'
              examples:
                example1:
                  summary: "Succesvolle response"
                  value:
                    onderzoeken:
                      - onderzoek_id: 42
                        organisatie_id: 1
                        status: "actief"
                        beschikbaar: true
                        type_onderzoek_id: 2
                        titel: "Onderzoek naar gezondheid"
                        beschrijving: "Dit is een onderzoek"
                        plaats: "Rotterdam"
                        max_deelnemers: 100
                        min_leeftijd: 18
                        max_leeftijd: 65
                        begeleiders: "Dr. Jansen"
                        datum: "2025-04-01"
                        datum_tot: "2025-06-01"
                        beloning: "€50"
                        creatie_datum: "2025-03-19"
        '400':
          description: "API-sleutel ontbreekt in de aanvraag"
        '403':
          description: "Ongeldige API-sleutel"
        '500':
          description: "Interne serverfout"
      security:
        - ApiKeyAuth: []

  /example-array:
    get:
      summary: "Voorbeeld van een lijst"
      operationId: "exampleArray"
      responses:
        '200':
          description: "Lijst van items"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              examples:
                example1:
                  summary: "Voorbeeld array response"
                  value:
                    - "Item 1"
                    - "Item 2"
                    - "Item 3"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Api-Key
  schemas:
    Onderzoeksvraag:
      type: object
      properties:
        onderzoek_id:
          type: integer
        organisatie_id:
          type: integer
        status:
          type: string
        beschikbaar:
          type: boolean
        type_onderzoek_id:
          type: integer
        titel:
          type: string
        beschrijving:
          type: string
        plaats:
          type: string
        max_deelnemers:
          type: integer
        min_leeftijd:
          type: integer
        max_leeftijd:
          type: integer
        beloning:
          type: string
        datum:
          type: string
          format: date
        datum_tot:
          type: string
          format: date
        begeleiders:
          type: string
        creatie_datum:
          type: string
          format: date
    Organisatie:
      type: object
      properties:
        naam:
          type: string
        email:
          type: string
        telefoonnummer:
          type: string
        contactpersoon:
          type: string
        beschrijving:
          type: string
        website:
          type: string
        adres:
          type: string